kustomize: change folder layout to comply
In kustomize v2.0 and higher, the resources in an overlay cannot
refer to a path outside the root folder.
This broke the current layout.
To fix it, move the contents of the default folder in the
config folder.
This will also be used as a base for various overlays (GKE, EKS, etc)

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>