member_controller: now using scylla manager agent
There is Scylla Client implementation mostly borrowed and adapted
from the Scylla Manager code itself. The controller uses this client
for status checks and decommission. The previous implementation that
used the eminent Jolokia library is removed and direct interaction
with the JMX server does no longer occur. There is still nodetool
invocation as part of the scylla container lifecycle to ensure that
proper shutdowns are performed by executing "drain" before
terminating the instance.