Add a install-dependencies.sh script
This is similar to what is in seastar and scylla. I am pretty sure it
is incomplete, but it is easy to add more distros and dependencies.

This avoids all binary packages and I have been able to build the
operator on my raspberrypi :-)

I have selected the version in order to cause the smallest possible
change to go.mod in the following patch.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>