Provision autogenerated auth token for Scylla API
From now on Agent requires a proper authorization token for requests.
Token is provisioned by the Operator as a Secret for each ScyllaCluster.
If user already enabled authentication by filling it in optional Agent
config Secret, its value will be used, if not, a random autogenerated string will be saved.