Use PodIP when hostNetwork is enabled instead of member service IP
Currently the advertised IP for cluster member is the IP of the member service.
This IP is often not accessible outside of the kubernetes cluster which limits the interractions of clients
and other scylla node (such as cross dc cluster).
Member service will provide a new label scylla/IP which will store the PodIP when hostNetwork is enabled.
IP from this label is then used for common operations (replace, scale down) and from the sidecar operator to define listen, seed, rpc address

Co-authored-by: n.fraison <n.fraison@criteo.com>