Automated EKS and GKE performance tuning
We have to calculate CPU mask for CPUs handling IRQs and those dedicated for Scylla.
Because CPU allocation is known only after Scylla container is created, the perftune is being run during sidecar startup.

In addition for GKE the disk write cache is disabled which greatly
improves overall performance.

New CRD field controls if tuning should be performed.
It is not always executed because it requires root priviledges and host PID namespace, and it may not work on dev environments.