Fix default anti-affinity rules for scylla-operator
For the moment 'scylla-operator' pods have default pod anti-affinity
rules set to be scheduled not with "webhook-server" ones.
The same vice-versa. It makes no sense because pods of the same type
may be scheduled on the same node in such a case.
It violates the HA idea.
So, fix it by switching labels between these affinity rules making
each pod of a type be scheduled on a node where there are no pods of
such a type.