sidecar: fix parallel racks creation
If multiple racks are present in the initial ScyllaCluster CRD multiple StatefulSet resources will be created. This results in multiple Pods being created in parallel.
Since the logic for the first node only triggers when there are no other Pods it will instead now create a circular dependency between the StatefulSets as seed nodes.