Fix flaky E2E test checking in-cluster shardawareness
Both on CI and on local environments traffic between test binary and
Scylla Pods is NATed. Previous implementation gave false positives
because we are low on resources (we use just 2 shards) so probability of
hitting expected shard was high even for random assignment.
It wasn't spotted because message about wrong shard assignment is
printed out by the driver to stdout and no error is reported anywhere.

Gocql doesn't expose any way of checking if shardaware ports are used
successfully, hence I implemented simple driver within a test, able to
send initial packet and reading to which shard connection was established.

It also check in-cluster connectivity which is how clients are
connecting from, compared to from outside in previous appraoch.

Fixes #1028