Fix flakyness of replace node e2e
Full quorum check is sensitive to seeing nodes not being part of the
cluster, when such node is spotted it immediately returns an error.
In replace node e2e, it could happen that cluster become ready before
replaced node was acknowledged by the others.
In this case, full quorum check returned an error, because it saw a IP
not bound to any cluster Service.

To fix it, test waits until cluster is rolled out, and replaced node is
seen by the other one, before the full quorum check is validated.