Fix Orphaned PV E2E
Test logic relied on fact that minikube storage provisioner did not set
any NodeAffinity for provisioned PVs. This allowed logic to trigger
cleanup logic by adding a NodeAffinity pointing to not existing Node.
When proper local volume proviosiner is used, and NodeAffinity is
configured for new PVs, test no longer have option to trigger cleanup
logic because NodeAffinity field is immutable.

To overcome this, the ScyllaCluster is using custom StorageClass not
backed by any provisioner running in the cluster.
ScyllaCluster is going to request PVC from that StorageClass, and the test
is going to request a clone of the original PVC
from the default StorageClass to get any storage. Then the bound PV is
rebounded to the original PVC
but with empty NodeAffinity. This allows the test to trigger the
orphaned PV cleanup logic.