Fix node setup image in EKS example
Image used by DaemonSet setting up the disks in EKS example got broken
when repository of 4.1 Scylla was removed.
In order to unstuck users using it, new version was released disabling
the removed repository.

Users which are affected by this should be aware that we are planning on
deprecating this DaemonSet in favor of upcoming automated disk setup.

Fixes #1218