Enhancement proposal for replacing nodes using their HostIDs
Currently, the Scylla Operator uses IP addresses of a node to identify
which node should be replaced. The issue is that depending on the
ScyllaCluster configuration, IP addresses might be ephemeral so it’s
error prone to use it as an identifier as they may change during when
the procedure is executed. Because there is a better stable alternative,
IP address approach became deprecated, hence we need to shift to it.

Details can be found in actual enhancement.