Add enhancment proposal for exposing ScyllaCluster
The proposed change aims to improve the accessibility of ScyllaDB nodes
by modifying the way they are exposed. By
implementing a switch to specify the type of IP addresses to use, such
as Pod IPs or LoadBalancer services, it would
allow reaching the nodes from external clients or other data centers.