Move resolution of `developerMode` and `additionalScyllaDBArguments` to Operator
Moves the resolution of developerMode and additionalScyllaDBArguments from the sidecar to the Operator.
This change addresses an issue where modifications to these arguments did not trigger a rolling restart of the ScyllaDB cluster.
By moving the logic to the Operator, any changes to these fields now correctly trigger a rolling restart as intended.

Since the sidecar no longer needs to fetch the ScyllaCluster resource for these parameters, it no longer requires permissions to perform GET operations on ScyllaCluster objects.
However, to ensure a safe rolling upgrade of the Operator, these permissions are not removed yet.