Disable ScyllaDBMonitoring tests in OpenShift E2E
This is because our current implementation of ScyllaDBMonitoring creates
a Prometheus stack that interferes with OpenShift's Prometheus pattern.
This problem is described in https://github.com/scylladb/scylla-operator/issues/2319.

Therefore, we do not expect ScyllaDBMonitoring in its current form to be usable
on OpenShift. Instead, we intend to implement one of the solutions listed in
a comment to the linked issue. Until then, we consider monitoring incompatible
with OpenShift.