Set ClientHealthcheckControllerAvailable condition status when RemoteKubernetesCluster healthcheck probes succeeds
Previously aggregated Available condition was set to True when there was no ClientHealthcheckControllerAvailable having False status.
When access to Kubernetes is broken and then restore, nothing removed this failed condition, hence RemoteKubernetesCluster was constantly not Available.
To fix that, healthcheck controller sets this condition according to state of given healthcheck iteration.