Remove the `vendor` setting from the dependabot config
There is inconsistency in the Dependabot docs - on one hand the example
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/controlling-dependencies-updated#updating-vendored-dependencies
allows a `vendor` value to be set, and the options reference https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#vendor-- expects it,

on the other hand there's a failing validation and a lot of public repos that happily use Dependabot without explicitly setting `vendor: true`.

This commit aims to keep the vendoring behavior enabled by implicit
autodetection by dependabot.