Make the Git references in docs point at the release branch instead of master
Before this PR, the branch part in the GitHub URLs was not correctly rendered, because it always pointed to "master".

The reason is that for all versions, `sphinx-multiversion` evaluates the `conf.py` originating from `master` and ignores the `conf.py` on the branch being rendered. This causes the `revision` value to always come from `master`, resulting in all links pointing to `master`. Which means that any change to the manifests can break all earlier minor versions, which is unacceptable.

This PR achieves 2 things
- replaces the incorrect `revision` field with `config.env.smv_current_version` which has the right branch name :tada:
- works around the fact that (in our setup) MyST's `substitution` plugin (for reasons beyond my comprehension) only allows "elaborate" Jinja templates (=such that are anything more than trivial resolution of values coming directly from `myst_substitutions`) outside of `code-block`s. We still want to render the code block, so we make the templates render a `code-block` without it being directly interpretable as a `code-block`, which would prevent proper evaluation of the Jinja expression.

I am **very** unproud of the style of this solution, but it achieves:
- correct links (finally!!)
- rich rendering of the code block

at the expense of hackiness that makes me feel bad about it. I think that a reasonable next step is to reach out to the maintainers of our docs stack for a better solution to replace this abomination.