Fix loop devices manipulations in Node Setup E2Es
When creating a loop device, it gets assigned `/dev/loop0`, `/dev/loop1` paths, and so on.
NodeConfig controller creates a symlink `/dev/loops/{name} -> /dev/loop0` for easier access.
Client container when reading the symlink under `/host/dev/loops/{name}` may get `/dev/loop0`
that accidentally always exists in the container and because of that tests can falsely succeed.
That is not the device on the host, but on the container. That explains why we were getting
`cannot open /host/dev/loops/disk: No such file or directory` - it happens when there are more
than 1 loop device created on the host, e.g., when our symlink resolves to `/dev/loop1`.
It will be the case because such a device doesn’t exist in the container.

We have manually resolve symlinks in the container and prepend `/host` to their paths when we
want to manipulate their filesystem.